Refactor transaction builder to solve for the complex linting issues
