+++ Sources/Transaction/TransactionBuilder.swift
@@ -27,6 +27,43 @@ extension TransactionBuilderError: CustomStringConvertible {
 }
 
 final class TransactionBuilder {
+    private let tombstoneBlockIndex: UInt64
+
+    private let ptr: OpaquePointer
+
+    private let memoBuilder: TxOutMemoBuilder
+
+    private init(
+        fee: Amount,
+        tombstoneBlockIndex: UInt64,
+        fogResolver: FogResolver = FogResolver(),
+        memoBuilder: TxOutMemoBuilder = DefaultMemoBuilder(),
+        blockVersion: BlockVersion
+    ) {
+        self.tombstoneBlockIndex = tombstoneBlockIndex
+        self.memoBuilder = memoBuilder
+        self.ptr = memoBuilder.withUnsafeOpaquePointer { memoBuilderPtr in
+            fogResolver.withUnsafeOpaquePointer { fogResolverPtr in
+                // Safety: mc_transaction_builder_create should never return nil.
+                withMcInfallible {
+                    mc_transaction_builder_create(
+                            fee.value,
+                            fee.tokenId.value,
+                            tombstoneBlockIndex,
+                            fogResolverPtr,
+                            memoBuilderPtr,
+                            blockVersion)
+                }
+            }
+        }
+    }
+
+    deinit {
+        mc_transaction_builder_free(ptr)
+    }
+}
+
+extension TransactionBuilder {
     static func build(
         inputs: [PreparedTxInput],
         accountKey: AccountKey,
@@ -68,26 +105,27 @@ final class TransactionBuilder {
         rng: (@convention(c) (UnsafeMutableRawPointer?) -> UInt64)? = securityRNG,
         rngContext: Any? = nil
     ) -> Result<PendingSinglePayloadTransaction, TransactionBuilderError> {
-        positiveRemainingAmount(inputValues: inputs.map { $0.knownTxOut.value }, fee: fee)
-            .map { outputAmount in
-                PossibleTransaction([TransactionOutput(recipient, outputAmount)], nil)
+        Math.positiveRemainingAmount(
+            inputValues: inputs.map { $0.knownTxOut.value },
+            fee: fee
+        ).map { outputAmount in
+            PossibleTransaction([TransactionOutput(recipient, outputAmount)], nil)
+        }.flatMap { possibleTransaction in
+            build(
+                inputs: inputs,
+                accountKey: accountKey,
+                possibleTransaction: possibleTransaction,
+                memoType: memoType,
+                fee: fee,
+                tombstoneBlockIndex: tombstoneBlockIndex,
+                fogResolver: fogResolver,
+                blockVersion: blockVersion,
+                rng: rng,
+                rngContext: rngContext
+            ).map { pendingTransaction in
+                pendingTransaction.singlePayload
             }
-            .flatMap { possibleTransaction in
-                build(
-                    inputs: inputs,
-                    accountKey: accountKey,
-                    possibleTransaction: possibleTransaction,
-                    memoType: memoType,
-                    fee: fee,
-                    tombstoneBlockIndex: tombstoneBlockIndex,
-                    fogResolver: fogResolver,
-                    blockVersion: blockVersion,
-                    rng: rng,
-                    rngContext: rngContext
-                ).map { pendingTransaction in
-                    pendingTransaction.singlePayload
-                }
-            }
+        }
     }
 
     static func build(
@@ -102,7 +140,7 @@ final class TransactionBuilder {
         rng: (@convention(c) (UnsafeMutableRawPointer?) -> UInt64)? = securityRNG,
         rngContext: Any? = nil
     ) -> Result<PendingTransaction, TransactionBuilderError> {
-        outputsAddingChangeOutputIfNeeded(
+        outputsAddingChangeOutput(
             inputs: inputs,
             outputs: outputs,
             fee: fee
@@ -133,13 +171,7 @@ final class TransactionBuilder {
         rng: (@convention(c) (UnsafeMutableRawPointer?) -> UInt64)? = securityRNG,
         rngContext: Any? = nil
     ) -> Result<PendingTransaction, TransactionBuilderError> {
-        let outputs = possibleTransaction.outputs
-        let changeAmount = possibleTransaction.changeAmount
-        let allValues = (outputs.map { $0.amount.value } + [fee.value, changeAmount?.value ?? 0])
-        guard UInt64.safeCompare(
-                sumOfValues: inputs.map { $0.knownTxOut.value },
-                isEqualToSumOfValues: allValues)
-        else {
+        guard Math.totalOutlayCheck(for: possibleTransaction, fee: fee, inputs: inputs) else {
             return .failure(.invalidInput("Input values != output values + fee"))
         }
 
@@ -158,7 +190,7 @@ final class TransactionBuilder {
             }
         }
 
-        let payloadContexts = outputs.map { output in
+        let payloadContexts = possibleTransaction.outputs.map { output in
             builder.addOutput(
                 publicAddress: output.recipient,
                 amount: output.amount.value,
@@ -167,23 +199,11 @@ final class TransactionBuilder {
             )
         }
 
-        let changeContext: Result<TxOutContext, TransactionBuilderError>
-        switch blockVersion {
-        case .legacy:
-            // Clients built for BlockVersion == 0 (.legacy) will have trouble finding txOuts
-            // on the new change subaddress (max - 1), so we will emulate legacy behavior.
-            changeContext = builder.addOutput(
-                publicAddress: accountKey.publicAddress,
-                amount: changeAmount?.value ?? 0,
-                rng: rng,
-                rngContext: rngContext)
-        default:
-            changeContext = builder.addChangeOutput(
-                accountKey: accountKey,
-                amount: changeAmount?.value ?? 0,
-                rng: rng,
-                rngContext: rngContext)
-        }
+        let changeContext = changeContext(
+            blockVersion: blockVersion,
+            accountKey: accountKey,
+            builder: builder,
+            changeAmount: possibleTransaction.changeAmount)
 
         return payloadContexts.collectResult().flatMap { payloadContexts in
             changeContext.flatMap { changeContext in
@@ -194,6 +214,32 @@ final class TransactionBuilder {
                         changeTxOutContext: changeContext)
                 }
             }
+        }
+    }
+
+    private static func changeContext(
+        blockVersion: BlockVersion,
+        accountKey: AccountKey,
+        builder: TransactionBuilder,
+        changeAmount: PositiveUInt64?,
+        rng: (@convention(c) (UnsafeMutableRawPointer?) -> UInt64)? = securityRNG,
+        rngContext: Any? = nil
+    ) -> Result<TxOutContext, TransactionBuilderError> {
+        switch blockVersion {
+        case .legacy:
+            // Clients built for BlockVersion == 0 (.legacy) will have trouble finding txOuts
+            // on the new change subaddress (max - 1), so we will emulate legacy behavior.
+            return builder.addOutput(
+                publicAddress: accountKey.publicAddress,
+                amount: changeAmount?.value ?? 0,
+                rng: rng,
+                rngContext: rngContext)
+        default:
+            return builder.addChangeOutput(
+                accountKey: accountKey,
+                amount: changeAmount?.value ?? 0,
+                rng: rng,
+                rngContext: rngContext)
         }
     }
 
@@ -237,99 +283,23 @@ final class TransactionBuilder {
             rngContext: rngContext)
     }
 
-    // TODO - Should we always add change output for BlockVersion == 0 ?
-    private static func outputsAddingChangeOutputIfNeeded(
+    private static func outputsAddingChangeOutput(
         inputs: [PreparedTxInput],
         outputs: [TransactionOutput],
         fee: Amount
     ) -> Result<PossibleTransaction, TransactionBuilderError> {
-        remainingAmount(
+        Math.remainingAmount(
             inputValues: inputs.map { $0.knownTxOut.value },
             outputValues: outputs.map { $0.amount.value },
             fee: fee
-        ).map { remainingAmount in
+        )
+        .map { remainingAmount in
             PossibleTransaction(outputs, PositiveUInt64(remainingAmount))
         }
     }
+}
 
-    private static func remainingAmount(inputValues: [UInt64], outputValues: [UInt64], fee: Amount)
-        -> Result<UInt64, TransactionBuilderError>
-    {
-        guard UInt64.safeCompare(
-                sumOfValues: inputValues,
-                isGreaterThanOrEqualToSumOfValues: outputValues + [fee.value])
-        else {
-            return .failure(.invalidInput("Total input amount < total output amount + fee"))
-        }
-
-        guard let remainingAmount = UInt64.safeSubtract(
-                sumOfValues: inputValues,
-                minusSumOfValues: outputValues + [fee.value])
-        else {
-            return .failure(.invalidInput("Change amount overflows UInt64"))
-        }
-
-        return .success(remainingAmount)
-    }
-
-    private static func positiveRemainingAmount(inputValues: [UInt64], fee: Amount)
-        -> Result<PositiveUInt64, TransactionBuilderError>
-    {
-        guard UInt64.safeCompare(sumOfValues: inputValues, isGreaterThanValue: fee.value) else {
-            return .failure(.invalidInput("Total input amount <= fee"))
-        }
-
-        guard let remainingAmount = UInt64.safeSubtract(
-                sumOfValues: inputValues,
-                minusValue: fee.value)
-        else {
-            return .failure(.invalidInput("Change amount overflows UInt64"))
-        }
-
-        guard let positiveRemainingAmount = PositiveUInt64(remainingAmount) else {
-            // This condition should be redundant with the first check, but we throw an error
-            // anyway, rather than calling fatalError.
-            return .failure(.invalidInput("Total input amount == fee"))
-        }
-
-        return .success(positiveRemainingAmount)
-    }
-
-    private let tombstoneBlockIndex: UInt64
-
-    private let ptr: OpaquePointer
-
-    private let memoBuilder: TxOutMemoBuilder
-
-    private init(
-        fee: Amount,
-        tombstoneBlockIndex: UInt64,
-        fogResolver: FogResolver = FogResolver(),
-        memoBuilder: TxOutMemoBuilder = DefaultMemoBuilder(),
-        blockVersion: BlockVersion
-    ) {
-        self.tombstoneBlockIndex = tombstoneBlockIndex
-        self.memoBuilder = memoBuilder
-        self.ptr = memoBuilder.withUnsafeOpaquePointer { memoBuilderPtr in
-            fogResolver.withUnsafeOpaquePointer { fogResolverPtr in
-                // Safety: mc_transaction_builder_create should never return nil.
-                withMcInfallible {
-                    mc_transaction_builder_create(
-                            fee.value,
-                            fee.tokenId.value,
-                            tombstoneBlockIndex,
-                            fogResolverPtr,
-                            memoBuilderPtr,
-                            blockVersion)
-                }
-            }
-        }
-    }
-
-    deinit {
-        mc_transaction_builder_free(ptr)
-    }
-
+extension TransactionBuilder {
     private func addInput(preparedTxInput: PreparedTxInput, accountKey: AccountKey)
         -> Result<(), TransactionBuilderError>
     {
@@ -348,31 +318,11 @@ final class TransactionBuilder {
         viewPrivateKey: RistrettoPrivate,
         subaddressSpendPrivateKey: RistrettoPrivate
     ) -> Result<(), TransactionBuilderError> {
-        let ring = McTransactionBuilderRing(ring: preparedTxInput.ring)
-        return viewPrivateKey.asMcBuffer { viewPrivateKeyPtr in
-            subaddressSpendPrivateKey.asMcBuffer { subaddressSpendPrivateKeyPtr in
-                ring.withUnsafeOpaquePointer { ringPtr in
-                    withMcError { errorPtr in
-                        mc_transaction_builder_add_input(
-                            ptr,
-                            viewPrivateKeyPtr,
-                            subaddressSpendPrivateKeyPtr,
-                            preparedTxInput.realInputIndex,
-                            ringPtr,
-                            &errorPtr)
-                    }.mapError {
-                        switch $0.errorCode {
-                        case .invalidInput:
-                            return .invalidInput("\(redacting: $0.description)")
-                        default:
-                            // Safety: mc_transaction_builder_add_input should not throw
-                            // non-documented errors.
-                            logger.fatalError("Unhandled LibMobileCoin error: \(redacting: $0)")
-                        }
-                    }
-                }
-            }
-        }
+        TransactionBuilderUtils.addInput(
+            ptr: ptr,
+            preparedTxInput: preparedTxInput,
+            viewPrivateKey: viewPrivateKey,
+            subaddressSpendPrivateKey: subaddressSpendPrivateKey)
     }
 
     private func addOutput(
@@ -381,52 +331,13 @@ final class TransactionBuilder {
         rng: (@convention(c) (UnsafeMutableRawPointer?) -> UInt64)?,
         rngContext: Any?
     ) -> Result<TxOutContext, TransactionBuilderError> {
-        var confirmationNumberData = Data32()
-        var sharedSecretData = Data32()
-        return publicAddress.withUnsafeCStructPointer { publicAddressPtr in
-            withMcRngCallback(rng: rng, rngContext: rngContext) { rngCallbackPtr in
-                confirmationNumberData.asMcMutableBuffer { confirmationNumberPtr in
-                    sharedSecretData.asMcMutableBuffer { sharedSecretPtr in
-                        Data.make(withMcDataBytes: { errorPtr in
-                            mc_transaction_builder_add_output(
-                                ptr,
-                                amount,
-                                publicAddressPtr,
-                                rngCallbackPtr,
-                                confirmationNumberPtr,
-                                sharedSecretPtr,
-                                &errorPtr)
-                        }).mapError {
-                            switch $0.errorCode {
-                            case .invalidInput:
-                                return .invalidInput("\(redacting: $0.description)")
-                            case .attestationVerificationFailed:
-                                return .attestationVerificationFailed("\(redacting: $0.description)")
-                            default:
-                                // Safety: mc_transaction_builder_add_output should not throw
-                                // non-documented errors.
-                                logger.fatalError("Unhandled LibMobileCoin error: \(redacting: $0)")
-                            }
-                        }
-                    }
-                }
-            }
-        }.map { txOutData in
-            guard let txOut = TxOut(serializedData: txOutData) else {
-                // Safety: mc_transaction_builder_add_output should always return valid data on
-                // success.
-                logger.fatalError("mc_transaction_builder_add_output returned invalid data: " +
-                    "\(redacting: txOutData.base64EncodedString())")
-            }
-
-            let confirmationNumber = TxOutConfirmationNumber(confirmationNumberData)
-            let sharedSecret = RistrettoPublic(skippingValidation: sharedSecretData)
-            let receipt = Receipt(
-                txOut: txOut,
-                confirmationNumber: confirmationNumber,
-                tombstoneBlockIndex: tombstoneBlockIndex)
-            return TxOutContext(txOut, receipt, sharedSecret)
-        }
+        TransactionBuilderUtils.addOutput(
+            ptr: ptr,
+            tombstoneBlockIndex: tombstoneBlockIndex,
+            publicAddress: publicAddress,
+            amount: amount,
+            rng: rng,
+            rngContext: rngContext)
     }
 
     private func addChangeOutput(
@@ -435,118 +346,19 @@ final class TransactionBuilder {
         rng: (@convention(c) (UnsafeMutableRawPointer?) -> UInt64)?,
         rngContext: Any?
     ) -> Result<TxOutContext, TransactionBuilderError> {
-
-        var confirmationNumberData = Data32()
-        var sharedSecretData = Data32()
-
-        return McAccountKey.withUnsafePointer(
-            viewPrivateKey: accountKey.viewPrivateKey,
-            spendPrivateKey: accountKey.spendPrivateKey,
-            fogInfo: accountKey.fogInfo
-        ) { accountKeyPtr in
-                withMcRngCallback(rng: rng, rngContext: rngContext) { rngCallbackPtr in
-                    confirmationNumberData.asMcMutableBuffer { confirmationNumberPtr in
-                        sharedSecretData.asMcMutableBuffer { sharedSecretPtr in
-                            Data.make(withMcDataBytes: { errorPtr in
-                                mc_transaction_builder_add_change_output(
-                                    accountKeyPtr,
-                                    ptr,
-                                    amount,
-                                    rngCallbackPtr,
-                                    confirmationNumberPtr,
-                                    sharedSecretPtr,
-                                    &errorPtr)
-                            }).mapError {
-                                switch $0.errorCode {
-                                case .invalidInput:
-                                    return .invalidInput("\(redacting: $0.description)")
-                                case .attestationVerificationFailed:
-                                    return .attestationVerificationFailed("\(redacting: $0.description)")
-                                default:
-                                    // Safety: mc_transaction_builder_add_output should not throw
-                                    // non-documented errors.
-                                    logger.fatalError("Unhandled LibMobileCoin error: \(redacting: $0)")
-                                }
-                            }
-                        }
-                    }
-                }
-        }.map { txOutData in
-            guard let txOut = TxOut(serializedData: txOutData) else {
-                // Safety: mc_transaction_builder_add_output should always return valid data on
-                // success.
-                logger.fatalError("mc_transaction_builder_add_output returned invalid data: " +
-                    "\(redacting: txOutData.base64EncodedString())")
-            }
-
-            let confirmationNumber = TxOutConfirmationNumber(confirmationNumberData)
-            let sharedSecret = RistrettoPublic(skippingValidation: sharedSecretData)
-            let receipt = Receipt(
-                txOut: txOut,
-                confirmationNumber: confirmationNumber,
-                tombstoneBlockIndex: tombstoneBlockIndex)
-            return TxOutContext(txOut, receipt, sharedSecret)
-        }
+        TransactionBuilderUtils.addChangeOutput(
+            ptr: ptr,
+            tombstoneBlockIndex: tombstoneBlockIndex,
+            accountKey: accountKey,
+            amount: amount,
+            rng: rng,
+            rngContext: rngContext)
     }
 
     private func build(
         rng: (@convention(c) (UnsafeMutableRawPointer?) -> UInt64)?,
         rngContext: Any?
     ) -> Result<Transaction, TransactionBuilderError> {
-        withMcRngCallback(rng: rng, rngContext: rngContext) { rngCallbackPtr in
-            Data.make(withMcDataBytes: { errorPtr in
-                mc_transaction_builder_build(ptr, rngCallbackPtr, &errorPtr)
-            }).mapError {
-                switch $0.errorCode {
-                case .invalidInput:
-                    return .invalidInput("\(redacting: $0.description)")
-                default:
-                    // Safety: mc_transaction_builder_build should not throw non-documented errors.
-                    logger.fatalError("Unhandled LibMobileCoin error: \(redacting: $0)")
-                }
-            }
-        }.map { txBytes in
-            guard let transaction = Transaction(serializedData: txBytes) else {
-                // Safety: mc_transaction_builder_build should always return valid data on success.
-                logger.fatalError("mc_transaction_builder_build returned invalid data: " +
-                    "\(redacting: txBytes.base64EncodedString())")
-            }
-            return transaction
-        }
-    }
-}
-
-private final class McTransactionBuilderRing {
-    private let ptr: OpaquePointer
-
-    init(ring: [(TxOut, TxOutMembershipProof)]) {
-        // Safety: mc_transaction_builder_ring_create should never return nil.
-        self.ptr = withMcInfallible(mc_transaction_builder_ring_create)
-
-        for (txOut, membershipProof) in ring {
-            addElement(txOut: txOut, membershipProof: membershipProof)
-        }
-    }
-
-    deinit {
-        mc_transaction_builder_ring_free(ptr)
-    }
-
-    func addElement(txOut: TxOut, membershipProof: TxOutMembershipProof) {
-        txOut.serializedData.asMcBuffer { txOutBytesPtr in
-            membershipProof.serializedData.asMcBuffer { membershipProofDataPtr in
-                // Safety: mc_transaction_builder_ring_add_element should never return nil.
-                withMcInfallible {
-                    mc_transaction_builder_ring_add_element(
-                        ptr,
-                        txOutBytesPtr,
-                        membershipProofDataPtr)
-                }
-            }
-        }
-    }
-
-    func withUnsafeOpaquePointer<R>(_ body: (OpaquePointer) throws -> R) rethrows -> R {
-        try body(ptr)
+        TransactionBuilderUtils.build(ptr: ptr, rng: rng, rngContext: rngContext)
     }
 }
